Grand value division: value::de
